primera división de main a data.js
