+++ src/index.html
@@ -17,7 +17,7 @@
       <ul>
         <li class="header"><a href="indexHistoria.html">Historia</a></li>
         <li class="header"><a href="index.html">Películas</a></li>
-        <li id= "clickTop" class="header">Top</li>
+        <li id= "clickTop" class="header"><a>Top 10</a></li>
       </ul>
     </nav>
     </header>
